Make menuReload actually reload the menu
Use menuReload in registerWebSocketListeners
