+++ packages/components/src/index.ts
@@ -687,7 +687,6 @@ import {
     isSampleStatusEnabled,
     isSharedDefinition,
     isWorkflowEnabled,
-    registerWebSocketListeners,
     sampleManagerIsPrimaryApp,
     useMenuSectionConfigs,
     userCanDeletePublicPicklists,
@@ -708,6 +707,7 @@ import {
     menuInit,
     menuInvalidate,
     menuReload,
+    registerWebSocketListeners,
     serverNotificationInit,
     serverNotificationInvalidate,
     updateUser,
